requirements.txt
newsapi-python==0.2.7
openai
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
streamlit
scikit-learn==1.3.2
nltk==3.8.1
matplotlib==3.7.5
seaborn==0.13.0
transformers==4.36.2
torch==2.1.2
tqdm==4.66.1
spacy
textblob
yfinance
plotly==5.18.0
